Separated Fortify installation into its own command
